refactor: 가독성 개선
